Refactor to generic validateKeys
