+++ api/src/services/items.ts
@@ -21,7 +21,7 @@ import getASTFromQuery from '../utils/get-ast-from-query';
 import { AuthorizationService } from './authorization';
 import { ActivityService, RevisionsService } from './index';
 import { PayloadService } from './payload';
-import { validatePrimaryKeys } from '../utils/validate-primary-keys';
+import { validateKeys } from '../utils/validate-keys';
 
 export type QueryOptions = {
 	stripNonRequested?: boolean;
@@ -320,7 +320,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 	 */
 	async readOne(key: PrimaryKey, query: Query = {}, opts?: QueryOptions): Promise<Item> {
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, key);
+		validateKeys(this.schema, this.collection, primaryKeyField, key);
 
 		const filterWithKey = assign({}, query.filter, { [primaryKeyField]: { _eq: key } });
 		const queryWithKey = assign({}, query, { filter: filterWithKey });
@@ -339,7 +339,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 	 */
 	async readMany(keys: PrimaryKey[], query: Query = {}, opts?: QueryOptions): Promise<Item[]> {
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, keys);
+		validateKeys(this.schema, this.collection, primaryKeyField, keys);
 
 		const filterWithKey = { _and: [{ [primaryKeyField]: { _in: keys } }, query.filter ?? {}] };
 		const queryWithKey = assign({}, query, { filter: filterWithKey });
@@ -361,7 +361,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 		const keys = await this.getKeysByQuery(query);
 
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, keys);
+		validateKeys(this.schema, this.collection, primaryKeyField, keys);
 
 		return keys.length ? await this.updateMany(keys, data, opts) : [];
 	}
@@ -371,7 +371,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 	 */
 	async updateOne(key: PrimaryKey, data: Partial<Item>, opts?: MutationOptions): Promise<PrimaryKey> {
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, key);
+		validateKeys(this.schema, this.collection, primaryKeyField, key);
 
 		await this.updateMany([key], data, opts);
 		return key;
@@ -382,7 +382,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 	 */
 	async updateMany(keys: PrimaryKey[], data: Partial<Item>, opts?: MutationOptions): Promise<PrimaryKey[]> {
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, keys);
+		validateKeys(this.schema, this.collection, primaryKeyField, keys);
 
 		const fields = Object.keys(this.schema.collections[this.collection].fields);
 		const aliases = Object.values(this.schema.collections[this.collection].fields)
@@ -557,7 +557,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 		const primaryKey: PrimaryKey | undefined = payload[primaryKeyField];
 
 		if (primaryKey) {
-			validatePrimaryKeys(this.schema, this.collection, primaryKeyField, primaryKey);
+			validateKeys(this.schema, this.collection, primaryKeyField, primaryKey);
 		}
 
 		const exists =
@@ -610,7 +610,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 		const keys = await this.getKeysByQuery(query);
 
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, keys);
+		validateKeys(this.schema, this.collection, primaryKeyField, keys);
 
 		return keys.length ? await this.deleteMany(keys, opts) : [];
 	}
@@ -620,7 +620,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 	 */
 	async deleteOne(key: PrimaryKey, opts?: MutationOptions): Promise<PrimaryKey> {
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, key);
+		validateKeys(this.schema, this.collection, primaryKeyField, key);
 
 		await this.deleteMany([key], opts);
 		return key;
@@ -631,7 +631,7 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 	 */
 	async deleteMany(keys: PrimaryKey[], opts?: MutationOptions): Promise<PrimaryKey[]> {
 		const primaryKeyField = this.schema.collections[this.collection].primary;
-		validatePrimaryKeys(this.schema, this.collection, primaryKeyField, keys);
+		validateKeys(this.schema, this.collection, primaryKeyField, keys);
 
 		if (this.accountability && this.accountability.admin !== true) {
 			const authorizationService = new AuthorizationService({
