+++ api/src/utils/validate-keys.ts
@@ -4,20 +4,20 @@ import { PrimaryKey } from '../types';
 import validateUUID from 'uuid-validate';
 
 /**
- * Validate primary keys based on its type
+ * Validate keys based on its type
  */
-export function validatePrimaryKeys(
+export function validateKeys(
 	schema: SchemaOverview,
 	collection: string,
-	primaryKeyField: string,
+	keyField: string,
 	keys: PrimaryKey | PrimaryKey[]
 ) {
 	if (Array.isArray(keys)) {
 		for (const key of keys) {
-			validatePrimaryKeys(schema, collection, primaryKeyField, key);
+			validateKeys(schema, collection, keyField, key);
 		}
 	} else {
-		const primaryKeyFieldType = schema.collections[collection].fields[primaryKeyField].type;
+		const primaryKeyFieldType = schema.collections[collection].fields[keyField].type;
 		if (primaryKeyFieldType === 'uuid' && !validateUUID(String(keys))) {
 			throw new ForbiddenException();
 		} else if (primaryKeyFieldType === 'integer' && !Number.isInteger(Number(keys))) {
